refactor(terminal): split TerminalInstanceService into focused helper modules
- Extract TerminalOffscreenManager for hidden container and offscreen slot lifecycle
- Extract TerminalLinkHandler for URL opening and browser panel behavior
- Extract TerminalResizeController for resize scheduling, locking, and dimension handling
- Extract TerminalRendererPolicy for tier management and backend streaming control
- Extract TerminalWakeManager for wake/restore operations with rate limiting
- Refactor TerminalInstanceService to delegate to helpers (1660 → 953 lines)
- Preserve all public APIs and behavior unchanged
